% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjson-package.R
\docType{package}
\name{hjson-package}
\alias{hjson}
\alias{hjson-package}
\title{hjson: R Interface to HJson-cpp}
\description{
Provides function to read HJson files into R lists.
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
