% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_hjson}
\alias{read_hjson}
\title{Read HJson
Reads an HJson file into an R List}
\usage{
read_hjson(path, ...)
}
\arguments{
\item{path}{path to hjson file}

\item{...}{params passed to jsonlite::fromJSON}
}
\description{
Read HJson
Reads an HJson file into an R List
}
